+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `scene.create_terrain()` now auto-computes normals from the heightmap gradients if `normals` is omitted.
+- `scene.create_terrain()` now accepts `grid_res` to auto-infer spatial `x_lim`/`y_lim` constraints instead of requiring manual definition.
+
 ## [3.5] - 2026-07-28
 
 ### Added